Improve module augmentation in `mention` plugin
