+++ packages/ckeditor5-mention/src/index.ts
@@ -10,3 +10,8 @@
 export { default as Mention } from './mention';
 export { default as MentionEditing } from './mentionediting';
 export { default as MentionUI } from './mentionui';
+
+export type { MentionConfig } from './mentionconfig';
+export type { default as MentionCommand } from './mentioncommand';
+
+import './augmentation';
